fix: PrivateRoute => SRP원칙으로 추상화
